+++ qcal/__init__.py
@@ -1,12 +1,15 @@
 """Quantum Calibration (qcal) package."""
 
-__version__ = "3.1.1"
+__version__ = "4.0.0"
 
 from qcal.circuit import *
 from qcal.compilation.compiler import Compiler
 from qcal.config import Config
-from qcal.gate.gate import Gate
-from qcal.gate.single_qubit import *
-from qcal.gate.two_qubit import *
+from qcal.gates.gate import Gate
+from qcal.gates.single_qubit import *
+from qcal.gates.single_qutrit import *
+from qcal.gates.two_qubit import *
+from qcal.gates.two_qutrit import *
 from qcal.results import Results
+from qcal.simulation import DensityMatrixSimulator, StateVectorSimulator
 from qcal.utils import *