+++ qcal/backend/__init__.py
@@ -1,1 +1,1 @@
-# from .qubic import qubic_transpiler
+from .emulator import Emulator